backend/requirements.txt
# ── Web Framework ───────────────────────────────────────────────────────────
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
websockets>=12,<16
python-multipart>=0.0.9

# ── Config & Validation ─────────────────────────────────────────────────────
pydantic>=2.6,<3.0
pydantic-settings>=2.6,<3.0
pydantic[email]>=2.6

# ── Authentication ──────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0
bcrypt<4.0

# ── Database Drivers ────────────────────────────────────────────────────────
# PostgreSQL async
asyncpg>=0.29,<1.0
sqlalchemy[asyncio]>=2.0,<3.0
alembic>=1.13,<2.0
geoalchemy2>=0.15,<1.0
psycopg2-binary>=2.9,<3.0

# MongoDB async
motor>=3.4,<4.0
pymongo>=4.7,<5.0

# Elasticsearch
elasticsearch[async]>=8.12,<9.0

# Redis async
redis[hiredis]>=5.0,<6.0

# ── Message Queue ───────────────────────────────────────────────────────────
confluent-kafka>=2.4,<3.0
celery[redis]>=5.3,<6.0

# ── NLP Pipeline ────────────────────────────────────────────────────────────
# spaCy is installed separately (models downloaded in Dockerfile)
spacy>=3.7,<4.0
langdetect>=1.0,<2.0
# HuggingFace (models downloaded on first use; cached in volume)
transformers>=4.40,<5.0
torch>=2.2; sys_platform != "darwin" or platform_machine != "arm64"
torch>=2.2; sys_platform == "darwin" and platform_machine == "arm64"
sentencepiece>=0.2,<1.0
huggingface-hub>=0.23,<1.0

# ── ML / Prediction ─────────────────────────────────────────────────────────
scikit-learn>=1.4,<2.0
numpy>=1.26,<2.0
pandas>=2.2,<3.0
scipy>=1.13,<2.0
xgboost>=2.0,<3.0
prophet>=1.1,<2.0
mlflow>=2.12,<3.0

# ── LangChain + Claude ───────────────────────────────────────────────────────
langchain>=0.2,<1.0
langchain-anthropic>=0.1,<1.0
anthropic>=0.25,<1.0

# ── Notifications ───────────────────────────────────────────────────────────
fastapi-mail>=1.4,<2.0
twilio>=9.0,<10.0
httpx>=0.28,<1.0
aiofiles>=23.2,<25.0

# ── Utilities ───────────────────────────────────────────────────────────────
email-validator>=2.2,<3.0
python-dateutil>=2.9,<3.0
pytz>=2024.1

# ── Monitoring ──────────────────────────────────────────────────────────────
prometheus-fastapi-instrumentator>=7.0,<8.0

# ── Testing ─────────────────────────────────────────────────────────────────
pytest>=8.3,<9.0
pytest-asyncio>=0.23,<1.0
httpx>=0.28,<1.0
